+++ application/views/template.php
@@ -6,6 +6,9 @@
         <meta HTTP-EQUIV="Content-Type" CONTENT="text/html; charset=UTF-8"/>
         <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
         <link rel="stylesheet" type="text/css" href="/assets/css/default.css"/>
+        <link rel="stylesheet" type="text/css" href="/assets/css/bootstrap.min.css"/>
+        <script src="/assets/js/jquery-3.1.1.slim.min.js"></script>
+        <script src="/assets/js/bootstrap.min.js"></script>
 	</head>
 	<body>
         <div id="container">